+++ src/components/NavigationBar/index.js
@@ -1,19 +1,25 @@
 import { h } from 'preact'
 import classNames from 'classnames'
+import { compose } from '../utils/func'
+import { withFullScreenState } from '../FullScreen'
 import style from './style.css'
-import {preventDefaultOnClick} from '../utils'
+import {preventDefaultOnClick, isDesktop} from '../utils'
+import {localised} from '../../locales'
 
-const NavigationBar = ({back, i18n, disabled, isFullScreen, className}) =>
+const NavigationBar = ({back, translate, disabled, isFullScreen, className}) =>
   <div className={classNames(className, style.navigation, {
     [style.fullScreenNav]: isFullScreen
   })}>
-    <button href='#' className={classNames(style.back, {[style.disabled]: disabled})}
+    <button href='#' className={classNames(style.back, {[style.disabled]: disabled, [style.backHoverDesktop]: isDesktop})}
       onClick={preventDefaultOnClick(back)}>
         <span className={style.iconBack} />
         <span className={style.label}>
-          {i18n.t('back')}
+          {translate('back')}
         </span>
     </button>
  </div>
 
-export default NavigationBar
+export default compose(
+  withFullScreenState,
+  localised
+)(NavigationBar)