+++ src/components/Router/StepComponentMap.js
@@ -2,15 +2,16 @@ import { h } from 'preact'
 
 import Welcome from '../Welcome'
 import {SelectPoADocument, SelectIdentityDocument} from '../Select'
-import {FrontDocumentCapture, BackDocumentCapture, FaceCapture, LivenessCapture} from '../Capture'
-import {DocumentFrontConfirm, DocumentBackConfirm, FaceConfirm, LivenessConfirm} from '../Confirm'
+import {FrontDocumentCapture, BackDocumentCapture, SelfieCapture, VideoCapture } from '../Capture'
+import {DocumentFrontConfirm, DocumentBackConfirm, SelfieConfirm, VideoConfirm} from '../Confirm'
 import Complete from '../Complete'
 import MobileFlow from '../crossDevice/MobileFlow'
 import CrossDeviceLink from '../crossDevice/CrossDeviceLink'
 import ClientSuccess from '../crossDevice/ClientSuccess'
 import CrossDeviceIntro from '../crossDevice/Intro'
-import LivenessIntro from '../Liveness/Intro'
-import PoADocumentCapture, { PoAIntro } from '../ProofOfAddress'
+import VideoIntro from '../Video/Intro'
+import { includes } from '../utils/array'
+import { PoACapture, PoAIntro, PoAGuidance } from '../ProofOfAddress'
 
 export const componentsList = ({flow, documentType, steps, mobileFlow}) => {
   const captureSteps = mobileFlow ? clientCaptureSteps(steps) : steps
@@ -26,7 +27,7 @@ const hasCompleteStep = (steps) => steps.some(isComplete)
 const clientCaptureSteps = (steps) =>
   hasCompleteStep(steps) ? steps : [...steps, {type: 'complete'}]
 
-const shouldUseLiveness = steps => {
+const shouldUseVideo = steps => {
   const { options: faceOptions } = Array.find(steps, ({ type }) => type === 'face') || {}
   return (faceOptions || {}).requestedVariant === 'video' && window.MediaRecorder
 }
@@ -36,11 +37,11 @@ const captureStepsComponents = (documentType, mobileFlow, steps) => {
 
   return {
     welcome: () => [Welcome],
-    face: () => shouldUseLiveness(steps) ?
-        [LivenessIntro, LivenessCapture, LivenessConfirm] :
-        [FaceCapture, FaceConfirm],
+    face: () => shouldUseVideo(steps) ?
+        [VideoIntro, VideoCapture, VideoConfirm] :
+        [SelfieCapture, SelfieConfirm],
     document: () => createIdentityDocumentComponents(documentType),
-    poa: () => [PoAIntro, SelectPoADocument, PoADocumentCapture, DocumentFrontConfirm],
+    poa: () => [PoAIntro, SelectPoADocument, PoAGuidance, PoACapture, DocumentFrontConfirm],
     complete: () => complete
   }
 }
@@ -48,7 +49,7 @@ const captureStepsComponents = (documentType, mobileFlow, steps) => {
 const createIdentityDocumentComponents = (documentType) => {
   const double_sided_docs = ['driving_licence', 'national_identity_card']
   const frontDocumentFlow = [SelectIdentityDocument, FrontDocumentCapture, DocumentFrontConfirm]
-  if (Array.includes(double_sided_docs, documentType)) {
+  if (includes(double_sided_docs, documentType)) {
     return [...frontDocumentFlow, BackDocumentCapture, DocumentBackConfirm]
   }
   return frontDocumentFlow