+++ src/components/utils/array.js
@@ -4,3 +4,5 @@ export const wrapArray = maybeArray => Array.isArray(maybeArray) ? maybeArray : 
 
 export const flatten = (...values) =>
   values.reduce((accum, arr) => accum.concat(...wrapArray(arr)), [])
+
+export const includes = (arr, value) => arr.filter(item => value === item).length > 0