+++ src/core/store/reducers/globals.js
@@ -6,8 +6,8 @@ const initialState = {
   socket: null,
   sms: {number: null, valid: false},
   clientSuccess: false,
-  i18n: null,
-  termsAccepted: false
+  termsAccepted: false,
+  isFullScreen: false,
 }
 
 
@@ -27,6 +27,8 @@ export default function globals(state = initialState, action) {
       return {...state, mobileConnected: action.payload}
     case constants.ACCEPT_TERMS:
       return {...state, termsAccepted: true}
+    case constants.SET_FULL_SCREEN:
+      return {...state, isFullScreen: !!action.payload}
     default:
       return state
   }