+++ src/index.js
@@ -1,25 +1,31 @@
 import { h, render } from 'preact'
-import { Provider } from 'react-redux'
+import { Provider as ReduxProvider } from 'react-redux'
 import EventEmitter from 'eventemitter2'
+import countries from 'react-phone-number-input/modules/countries'
 
-import { store, actions, selectors } from './core'
+import { store, actions } from './core'
 import Modal from './components/Modal'
 import Router from './components/Router'
 import Tracker from './Tracker'
+import { LocaleProvider } from './locales'
+import {lowerCase, upperCase} from './components/utils/string'
+import {includes} from './components/utils/array'
 
 const events = new EventEmitter()
 
 Tracker.setUp()
 
-const ModalApp = ({ options:{ useModal, isModalOpen, onModalRequestClose, buttonId, ...otherOptions}, ...otherProps}) =>
-  <Modal {...{useModal, buttonId}} isOpen={isModalOpen} onRequestClose={onModalRequestClose}>
+const ModalApp = ({ options:{ useModal, isModalOpen, onModalRequestClose, ...otherOptions}, ...otherProps}) =>
+  <Modal useModal={useModal} isOpen={isModalOpen} onRequestClose={onModalRequestClose}>
     <Router options={otherOptions} {...otherProps}/>
   </Modal>
 
 const Container = ({ options }) =>
-  <Provider store={store}>
-    <ModalApp options={options}/>
-  </Provider>
+  <ReduxProvider store={store}>
+    <LocaleProvider language={options.language}>
+      <ModalApp options={options} />
+    </LocaleProvider>
+  </ReduxProvider>
 
 /**
  * Renders the Onfido component
@@ -48,7 +54,6 @@ const noOp = ()=>{}
 
 const defaults = {
   token: 'some token',
-  buttonId: 'onfido-button',
   containerId: 'onfido-mount',
   onComplete: noOp
 }
@@ -56,8 +61,9 @@ const defaults = {
 const isStep = val => typeof val === 'object'
 const formatStep = typeOrStep => isStep(typeOrStep) ?  typeOrStep : {type:typeOrStep}
 
-const formatOptions = ({steps, ...otherOptions}) => ({
+const formatOptions = ({steps, smsNumberCountryCode, ...otherOptions}) => ({
   ...otherOptions,
+  smsNumberCountryCode: validateSmsCountryCode(smsNumberCountryCode),
   steps: (steps || ['welcome','document','face','complete']).map(formatStep)
 })
 
@@ -70,9 +76,22 @@ const deprecationWarnings = ({steps}) => {
   }
 }
 
+const isSMSCountryCodeValid = (smsNumberCountryCode) => {
+  const isCodeValid = includes(countries.map(([code]) => code), lowerCase(smsNumberCountryCode))
+  if (!isCodeValid) {
+    console.warn("`smsNumberCountryCode` must be a valid two-characters ISO Country Code. 'GB' will be used instead.")
+  }
+  return isCodeValid
+}
+
+const validateSmsCountryCode = (smsNumberCountryCode) => {
+  if (!smsNumberCountryCode) return 'GB'
+  return isSMSCountryCodeValid(smsNumberCountryCode) ? upperCase(smsNumberCountryCode) : 'GB'
+}
+
 Onfido.init = (opts) => {
   console.log("onfido_sdk_version", process.env.SDK_VERSION)
-  Tracker.track()
+  Tracker.install()
   const options = formatOptions({ ...defaults, ...opts, events })
   deprecationWarnings(options)
 
@@ -98,11 +117,12 @@ Onfido.init = (opts) => {
     },
 
     tearDown() {
-      const socket = selectors.socket(store.getState())
+      const { socket } = store.getState().globals
       socket && socket.close()
       actions.reset()
       events.removeAllListeners('complete')
       render(null, containerEl, this.element)
+      Tracker.uninstall()
     }
   }
 }