+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This change log file is based on best practices from [Keep a Changelog](http://k
 This project adheres to [Semantic Versioning](http://semver.org/). Breaking changes result in a different MAJOR version. UI changes that might break customizations on top of the SDK will be treated as breaking changes too.
 This project adheres to the Node [default version scheme](https://docs.npmjs.com/misc/semver).
 
+## [3.0.0] - 2018-10-30
+
+### Added
+- Public: Added support for default SMS number country code. The default country for the SMS number input can be customised by passing the `smsNumberCountryCode` option when the SDK is initialised. The value should be a 2-characters long ISO Country code string. If empty, the SMS number country code will default to `GB`.
+
+### Changed
+- Public: Remove support for `buttonId` initialization option
+- Internal: Woopra is imported without using eval
+
 ## [2.8.0] - 2018-09-20
 
 ### Changed
@@ -307,7 +316,8 @@ Install with `npm install onfido-sdk-ui@0.12.0-rc.1`
 
 
 [next-version]:
-https://github.com/onfido/onfido-sdk-ui/compare/2.8.0...development
+https://github.com/onfido/onfido-sdk-ui/compare/2.9.0...development
+[2.9.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.8.0...2.9.0
 [2.8.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.7.0...2.8.0
 [2.7.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.6.0...2.7.0
 [2.6.0]: https://github.com/onfido/onfido-sdk-ui/compare/2.5.0...2.6.0